add fonction

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: tgiraudo <tgiraudo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 19:06:23 by tgiraudo          #+#    #+#             */
-/*   Updated: 2022/11/16 16:45:57 by tgiraudo         ###   ########.fr       */
+/*   Updated: 2022/11/16 21:42:55 by tgiraudo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 int		ft_printf(const char *s, ...);
 int		ft_putchar(char c);
-int		ft_putnbr(int n);
+int		ft_putnbr(long int n);
 int		ft_putstr(char *s);
 int		ft_putnbr_base(long int nb, char *tab);
 int		print_adress(void *addr);
